[WIP] Windows: Fixes for echo in job_spec.lua
Several tests call the echo binary, but in windows echo is not an
executable but a shell command.

@equalsraf: looking for a suitable replacement for jobstart(['echo'].
If we don't mind calling the shell, then 'exit 0' or 'echo. > NUL; exit 0'
work in most cases, but the "can omit exit callback" test keeps failing, so
maybe there is an issue lurking there.